Refactor: move ExecutionState.kt to package
