+++ utbot-framework/src/main/kotlin/org/utbot/engine/DataClasses.kt
@@ -11,6 +11,7 @@ import org.utbot.engine.pc.UtIsExpression
 import org.utbot.engine.pc.UtTrue
 import org.utbot.engine.pc.mkAnd
 import org.utbot.engine.pc.mkOr
+import org.utbot.engine.state.ExecutionState
 import org.utbot.engine.symbolic.*
 import org.utbot.engine.types.TypeResolver
 import org.utbot.framework.plugin.api.FieldId
