+++ utbot-framework/src/main/kotlin/org/utbot/engine/selectors/PathSelector.kt
@@ -1,6 +1,6 @@
 package org.utbot.engine.selectors
 
-import org.utbot.engine.ExecutionState
+import org.utbot.engine.state.ExecutionState
 import org.utbot.engine.pc.UtSolverStatusKind
 
 /**
