+++ utbot-framework/src/main/kotlin/org/utbot/engine/selectors/nurs/DepthSelector.kt
@@ -1,6 +1,6 @@
 package org.utbot.engine.selectors.nurs
 
-import org.utbot.engine.ExecutionState
+import org.utbot.engine.state.ExecutionState
 import org.utbot.engine.selectors.strategies.ChoosingStrategy
 import org.utbot.engine.selectors.strategies.StoppingStrategy
