+++ utbot-framework/src/main/kotlin/org/utbot/engine/selectors/strategies/GeneratedTestCountingStatistics.kt
@@ -1,6 +1,6 @@
 package org.utbot.engine.selectors.strategies
 
-import org.utbot.engine.ExecutionState
+import org.utbot.engine.state.ExecutionState
 import org.utbot.engine.InterProceduralUnitGraph
 
 class GeneratedTestCountingStatistics(
