+++ utbot-framework/src/main/kotlin/org/utbot/engine/selectors/strategies/StatementsStatistics.kt
@@ -1,6 +1,6 @@
 package org.utbot.engine.selectors.strategies
 
-import org.utbot.engine.ExecutionState
+import org.utbot.engine.state.ExecutionState
 import org.utbot.engine.InterProceduralUnitGraph
 import soot.SootMethod
 import soot.jimple.Stmt
