+++ utbot-framework/src/main/kotlin/org/utbot/engine/symbolic/SymbolicState.kt
@@ -35,5 +35,4 @@ data class SymbolicState(
     fun stateForNestedMethod() = copy(
         memory = memory.memoryForNestedMethod()
     )
-
 }
